Fix non-null assertion on resolve() method call
There may be counter examples in some places, but in general the
resolve() method on any PsiReference can always return null. In some
cases this doesn't directly make sense, but incorrect code can be the
culprit of these types of issue a lot of the time: a reference to a
class that doesn't exist can never resolve. We have a few spots in our
code base where this mistake is made, so we definitely need to respect
the @Nullable marker on these methods.

Fixes #388