Fix AT features not working on EAP
Jetbrains changed the implementation of their Promise, so I needed a
different way to implement `srgMapNow`. This commit also fixes a few
issues regarding finding SRG and AT references in other McpModule
outside of the current module - something that happens frequently in any
multi-module project. Perhaps all of this can be generified to
understand module dependencies, but there are currently no checks
relating to module dependencies.

AT annotators are being run now, but for some reason aren't rendering,
which is still an open issue.

Closes #422