Update Forge project creation to use Maven metadata
Forge is no longer supporting the JSON API for Forge version data, so we
need to use Maven metadata instead.

This commit also removes the platform icon placement logic and instead
simply places it directly on the module it belongs to. While this
doesn't present as nicely to the user, the logic necessary to figure out
where to place the icon is every different situation is unmaintainable.

This commit also fixes a missing import in the Sponge main class
tempalte.

Fixes #505