Add guard against identical recursive case
I don't really know if this will fix the issue or not, but it doesn't
seem to be very common either. In any case, this will prevent the method
from being called with identical inputs twice in a row, which won't
solve more complex infinite recurision cases, but might help for simple
cases.

Fixes #480