Return null from fullQualifiedName when anonymous index fails
Any time something goes wrong when the class name can't be defined for
one reason or another, ClassNameResolutionFailedException is thrown, and
the public-facing APIs catch this exception and return null for failure.

The anonymous index method was instead throwing an IllegalStateException
on failure, which would propogate up instead of being caught.

Fixes #513