Remove complex translation call identification system
Rather than using a complex recursive, error-prone approach to
identifying translation calls, only direct calls to known functions are
now respected. This simplifies the handling a lot and ultimately leads
to more predictable results.

To re-enable the power of the old system, making translation functions
configurable is the next step.