Rework project creation system
This refactor aims to reduce the fragility of the current project
creation system, while increasing the maintainability as well. The
maintainability issue was caused by a variety of factors, but most
notably were:

 * Many unsafe non-null assertions (!!)
 * Tightly coupled buildsystem and project type code
 * Significant logic duplication between project types
 * Complex shared logic with confusing conditionals

This commit attempts to address these issues by building up project
creation plans as a composed collection of discrete shareable steps.
The steps allow sharing common logic between project types without
coupling them and helps to reduce code duplication. This system does
increase declaration duplication, as each project type declares a list
of steps individually for each project configuration (typically
different configurations are created for different build systems, but
that is not always the case), but I think duplicating declarations
aren't as bad as complex duplication of logic.

Since the steps for each project type are declared in a simple list the
project plan for each project configuration type is simpler to follow as
a linear execution of steps.