Partial update to 2020.1
Mostly nullability changes in the API that needs to be reflected in Kotlin source code.

Deprecation warnings are left for now, the only non-vital change I made was to convert components to listeners and postStartupActivities because they are deprecated.

Tests are failing for now, something related to facets not being available when the test cases are set up.