Rewrite translation handling (#613)
This is a complete rewrite of the translation (formerly i18n) component of the plugin. This was mostly motivated by the switch from `lang` to `json` files for translations and aims to make the plugin relatively agnostic with regards to translation format.

Key points of the new/improved implementation are as follows:
 * Translation files are indexed based on their locale. This allows very quick lookups and reverse searches ("where to find the translation for this key")
   * Index entries are collected both from JSON and old lang files
 * Only direct calls to (known) translation functions/constructors are respected
 * Intentions, actions etc. are agnostic of the underlying translation format

A side effect of the changes also is that sorting and "adding from default" are way more efficient than before.

The next step will be to add the option to configure what is considered a translation function. I'm mostly held up on the UI here. The configuration itself will be based on a versioned config system I've been developing.

I will also write an action for converting old lang files to the new JSON format.

Those two things will come in separate pull requests, since they're not fundamental to the underlying workings of the system.