Fix broken test and improve testing framework
This commit fixes the broken test in AmbiguousReferenceInspectionTest by
marking the Mixin class as `allowTreeAccessForFile` - allowing us to
access the AST even though the file is not in the in-memory editor. I
added this method call for all files added with the project builder,
since this is often something we'll want to do. I have no idea why this
caused inconsistent test failures, but adding that method does fix the
issue.

I also updated the test framework to use a fully in-memory project
directory setup, which goes well with the in-code project builder system
we use in this project. This allowed a lot of the complexity of the
project builder to be removed and should hopefully resolve any further
project structure setup or filesystem issues in the future.

The old MCLang (previously I18n) code was still using the old test
framework code, so I updated those to use the the new cleaner parser and
lexer test methods.

In the process of updating the MCLang code I stumbled on a significantly
cleaner fix for commenter tests that don't rely on reflection hacks and
can be accomplished entirely in-memory as well.