Add parameter for allowing AST access in test files
Most files shouldn't need AST access so it's still best to keep that
disabled. For the files where it might be necessary, there's a parameter
for that.

Switching the call from `getTypeElement().getType()` to just `getType()`
in `bytecode-util` is a possible optimization and shouldn't always need
AST access, but I'm not sure exactly how it works. I think stubs can be
available without the file being in the in-memory editor or something
like that. Either way, that should fix `appendDescriptor` from throwing
the AST access not allowed exception in the first place.