Restructure project
Move some stuff to a buildSrc Gradle plugin so it can be shared with a
new submodule (coming later) and remove the unncessary
com/demonwav/mcdev directory structure in the Kotlin source directory.

Kotlin does not require files to be in the full package directory like
Java, and the official style guide says to prefer not laying them out
like Java for non-shared modules. This project has existed longer than
that style guide, and that's a lot of files to move, so I just never
felt like move them until now.