Run all createMcpToSrg tasks in a single Gradle invocation
This significantly reduces the time and resources used by the
createMcpToSrg step of retrieving the SRG  maps the project is using.
Only a single daemon is used - potentially even re-used from the import
which just finished.

This also cleans up a lot of the code around that area and uses the
correct direct APIs to handle the sourcesets-as-modules option when
importing Gradle projects.

This commit also handles registering AT files with the Access
Transformer file type in FG3 based on the accessTransformers field in
the `minecraft` userdev extension.

Fixes #841