Fold invoker and accessor Mixin method calls to appear as the target method name
This folder attempts to make reading code which use invokers and
accessors a little cleaner by showing the user the method they are
actually calling in the end, rather than the large cast and weird
method name of the invoker.

The folds are configurable into 4 separate sections:
 1. Invoker method call fold
 2. Invoker cast expression fold
 3. Accessor method call fold
 4. Accessor cast expression fold

They are all enabled by default except for the accessor method call
fold, since some people may think the setter fold looks a little clunky.

This commit also improves the code folding options page by setting a
title for the fold groups, which presents them in a category together
and generally looks better.