Improve error reporter in various ways
This improves the previous attempt at linking stacktrace elements by
investigating the actual StackTraceElement objects rather than trying to
use regex to match stacktrace text. This change both makes the code much
simpler and much more robust.

The original code was written using regex because the Throwable object
you receive from the IdeaLoggingEvent isn't actually the error throwable
- it only contains the text for it. But I figured out that the actual
Throwable that was thrown can be retrieved from the data on the event,
if the data is LogMessage.

This commit also allows the new code to recognize existing issues using
the new format. I forgot to implement that last time so the new system
could only comment on existing issues made by the old error
notifications. While it would probably be better for the plugin to be
able to recognize both, since as time goes on the prevelance of the
older style (and their relevance) will decrease, so it seemed reasonable
to only support the current error message style - we can always mark new
issues as duplicates and close them ourselves.

Another change in this commit is the error reporter will only look at
the 300 most recent closed issues when looking for duplicates. It will
always look at all open issues. This is to prevent the error reporter
from getting too slow from too many issues - issues which have been
closed for years are probably not relevant to current problems.