Improve stack trace handling in issue reporter
This commit reverts the previous commit's statement on compatibility
with the previous error reporting system. By including the raw
stacktrace text in a <details> block the stacktrace is still compatible
with the previous error reporting system, and works as a fallback in
case an issue with the linking system messes up the stacktrace.

The formatted stacktrace is also now placed in a <pre><code> block
rather than just <pre>, as this causes GitHub's markdown renderer to
include a "copy to clipboard" button on the code block.