Fix gradle-intellij-plugin interaction on this project
Turns out the problem the who time was having the
`gradleToolingExtension` configuration extending the `idea`
configuration. I have no clue why that breaks everything, but removing
that fixes all of the issues with resolving classes in IntelliJ.

Turns out we didn't actually need that either. In order to get the
Gradle tooling extension sources to compile the dependency on Groovy
just needed to be re-added.