Exclude coroutine lib from IntelliJ dependency
The IntelliJ dependency does not provide sources for the coroutine
library, which can make it very frustrating to debug / view source code.
We specify our own coroutine library, but IntelliJ doesn't seem to
choose to use it over the IntelliJ library, I assume since the IntelliJ
library is defined first. Using an artifact transform we can filter out
the coroutine jar from the IntelliJ library, though, so the IDE will
only see the dependency we define.