Fix version json HTTP errors
The main issue is we were catching IOException after switching to fuel,
which wraps all exceptions in FuelError, so we weren't actually going to
the GitHub fallback.

We also saw this issue prior to switching to fuel, so as a failsafe to
the failsafe, I've added another fallback URL, hosted in a different
location, as a last-ditch effort to connect to something.